# jualdanbeli-STANDALONE/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: jualdanbeli Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/midtrans/token:
    post:
      operationId: getMidtransToken
      tags: [payments]
      summary: Get Midtrans Snap token for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId: { type: integer }
              required: [orderId]
      responses:
        "200":
          description: Snap token and redirect URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  redirectUrl: { type: string }
                  clientKey: { type: string }
                  isProduction: { type: boolean }
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reports:
    get:
      operationId: getReports
      tags: [reports]
      summary: List fraud/spam reports (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, investigating, resolved, dismissed]
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: submitReport
      tags: [reports]
      summary: Submit a fraud/spam report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{reportId}/resolve:
    patch:
      operationId: resolveReport
      tags: [reports]
      summary: Resolve or dismiss a report (admin only)
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportResolution"
      responses:
        "200":
          description: Report resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ─── ADMIN ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Report:
      type: object
      required: [id, reporterId, targetType, targetId, reason, status, createdAt]
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        targetType:
          type: string
          enum: [user, product, review, message]
        targetId:
          type: integer
        reason:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, investigating, resolved, dismissed]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [user, product, review, message]
        targetId:
          type: integer
        reason:
          type: string
          enum: [spam, fraud, fake_product, inappropriate, other]
        description:
          type: string
    ReportResolution:
      type: object
      required: [status, resolution]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        resolution:
          type: string

    # ─── ADMIN ─────────────────────────────────────────────────────
...